requirements.txt
Flask
flask-cors
pandas
scikit-learn
numpy
ollama
langchain
chromadb
langchain-text-splitters
langchain-community
langchain-core
langchain-chroma
langchain-ollama
langchain-huggingface
sentence-transformers==2.7.0
contourpy>=1.0.0,<2.0
gensim>=4.0.0,<5.0
numba>=0.56.0,<1.0
spaces>=0.19.4
omegaconf
huggingface-hub
opencv-python
gradio
gdown
python-dotenv
deep_translator
